Fix keys and ignore mod tap
Make the keymap usable again by adding in the ignoring of the mod tap,
as well as fixing the y key that was missed. Now all the layers have a
name on the screen as well so that should make it a bit easier now in
the beginning at least.